--- A2_Optimized_LifeCycle.php
@@ -1,237 +1,0 @@
-<?php
-/*
-    "WordPress Plugin Template" Copyright (C) 2014 Michael Simpson  (email : michael.d.simpson@gmail.com)
-
-    This file is part of WordPress Plugin Template for WordPress.
-
-    WordPress Plugin Template is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    WordPress Plugin Template is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with Contact Form to Database Extension.
-    If not, see http://www.gnu.org/licenses/gpl-3.0.html
-*/
-
-include_once('A2_Optimized_InstallIndicator.php');
-
-class A2_Optimized_LifeCycle extends A2_Optimized_InstallIndicator {
-
-    public function install() {
-
-        // Initialize Plugin Options
-        $this->initOptions();
-
-        // Initialize DB Tables used by the plugin
-        $this->installDatabaseTables();
-
-        // Other Plugin initialization - for the plugin writer to override as needed
-        $this->otherInstall();
-
-        // Record the installed version
-        $this->saveInstalledVersion();
-
-        // To avoid running install() more then once
-        $this->markAsInstalled();
-    }
-
-    public function uninstall() {
-        $this->otherUninstall();
-        $this->unInstallDatabaseTables();
-        $this->deleteSavedOptions();
-        $this->markAsUnInstalled();
-    }
-
-    /**
-     * Perform any version-upgrade activities prior to activation (e.g. database changes)
-     * @return void
-     */
-    public function upgrade() {
-    }
-
-    /**
-     * See: http://plugin.michael-simpson.com/?page_id=105
-     * @return void
-     */
-    public function activate() {
-    }
-
-    /**
-     * See: http://plugin.michael-simpson.com/?page_id=105
-     * @return void
-     */
-    public function deactivate() {
-    }
-
-    /**
-     * See: http://plugin.michael-simpson.com/?page_id=31
-     * @return void
-     */
-    protected function initOptions() {
-    }
-
-    public function addActionsAndFilters() {
-    }
-
-    /**
-     * See: http://plugin.michael-simpson.com/?page_id=101
-     * Called by install() to create any database tables if needed.
-     * Best Practice:
-     * (1) Prefix all table names with $wpdb->prefix
-     * (2) make table names lower case only
-     * @return void
-     */
-    protected function installDatabaseTables() {
-    }
-
-    /**
-     * See: http://plugin.michael-simpson.com/?page_id=101
-     * Drop plugin-created tables on uninstall.
-     * @return void
-     */
-    protected function unInstallDatabaseTables() {
-    }
-
-    /**
-     * Override to add any additional actions to be done at install time
-     * See: http://plugin.michael-simpson.com/?page_id=33
-     * @return void
-     */
-    protected function otherInstall() {
-    }
-
-    /**
-     * Override to add any additional actions to be done at uninstall time
-     * See: http://plugin.michael-simpson.com/?page_id=33
-     * @return void
-     */
-    protected function otherUninstall() {
-    }
-
-    /**
-     * Puts the configuration page in the Plugins menu by default.
-     * Override to put it elsewhere or create a set of submenus
-     * Override with an empty implementation if you don't want a configuration page
-     * @return void
-     */
-    public function addSettingsSubMenuPage() {
-        $this->addSettingsSubMenuPageToMenu();
-    }
-
-
-    protected function requireExtraPluginFiles() {
-        //require_once(ABSPATH . 'wp-includes/pluggable.php');
-        require_once(ABSPATH . 'wp-admin/includes/plugin.php');
-    }
-
-    /**
-     * @return string Slug name for the URL to the Setting page
-     * (i.e. the page for setting options)
-     */
-    protected function getSettingsSlug() {
-        return get_class($this) . '_admin';
-    }
-    
-	public function addAdminBar(){
-		$this->requireExtraPluginFiles();
-		global $wp_admin_bar;
-
-		if(current_user_can('manage_options')){
-			$wp_admin_bar->add_node(array(
-				'id'    => 'a2-optimized-admin-bar',
-				'title' => 'A2 Optimized',
-				'href'  => admin_url('admin.php?page='.$this->getSettingsSlug())
-			));
-		}
-	}
-
-
-	public function addLockedEditor(){
-		$this->requireExtraPluginFiles();
-		add_theme_page( '<span style="color:red !important">Editor Locked</span>', '<span style="color:red !important">Editor Locked</span>', 'manage_options', 'editor-locked', array(&$this,'settingsPage') );
-	}
-	
-
-    protected function addSettingsSubMenuPageToPluginsMenu() {
-        $this->requireExtraPluginFiles();
-        $displayName = $this->getPluginDisplayName();
-        add_submenu_page('plugins.php',
-                         $displayName,
-                         $displayName,
-                         'manage_options',
-                         $this->getSettingsSlug(),
-                         array(&$this, 'settingsPage'));
-    }
-    
-    protected function addSettingsSubMenuPageToDashboard() {
-        $this->requireExtraPluginFiles();
-        $displayName = $this->getPluginDisplayName();
-        add_dashboard_page($displayName,
-                         $displayName,
-                         'manage_options',
-                         $this->getSettingsSlug(),
-                         array(&$this, 'settingsPage'));
-    }
-    
-    
-    protected function addSettingsSubMenuPageToMenu() {
-        $this->requireExtraPluginFiles();
-        $displayName = $this->getPluginDisplayName();
-        add_menu_page(   $displayName,
-                         $displayName,
-                         'manage_options',
-                         $this->getSettingsSlug(),
-                         array(&$this, 'settingsPage'),null,3.14159265359);
-    }
-
-
-    protected function addSettingsSubMenuPageToSettingsMenu() {
-        $this->requireExtraPluginFiles();
-        $displayName = $this->getPluginDisplayName();
-        add_options_page($displayName,
-                         $displayName,
-                         'manage_options',
-                         $this->getSettingsSlug(),
-                         array(&$this, 'settingsPage'));
-    }
-
-    /**
-     * @param  $name string name of a database table
-     * @return string input prefixed with the WordPress DB table prefix
-     * plus the prefix for this plugin (lower-cased) to avoid table name collisions.
-     * The plugin prefix is lower-cases as a best practice that all DB table names are lower case to
-     * avoid issues on some platforms
-     */
-    protected function prefixTableName($name) {
-        global $wpdb;
-        return $wpdb->prefix .  strtolower($this->prefix($name));
-    }
-
-
-    /**
-     * Convenience function for creating AJAX URLs.
-     *
-     * @param $actionName string the name of the ajax action registered in a call like
-     * add_action('wp_ajax_actionName', array(&$this, 'functionName'));
-     *     and/or
-     * add_action('wp_ajax_nopriv_actionName', array(&$this, 'functionName'));
-     *
-     * If have an additional parameters to add to the Ajax call, e.g. an "id" parameter,
-     * you could call this function and append to the returned string like:
-     *    $url = $this->getAjaxUrl('myaction&id=') . urlencode($id);
-     * or more complex:
-     *    $url = sprintf($this->getAjaxUrl('myaction&id=%s&var2=%s&var3=%s'), urlencode($id), urlencode($var2), urlencode($var3));
-     *
-     * @return string URL that can be used in a web page to make an Ajax call to $this->functionName
-     */
-    public function getAjaxUrl($actionName) {
-        return admin_url('admin-ajax.php') . '?action=' . $actionName;
-    }
-
-}
