+++ A2_Optimized_Plugin.php
@@ -1,11 +1,40 @@
 <?php
-include_once('A2_Optimized_LifeCycle.php');
+include_once('A2_Optimized_OptionsManager.php');
 
 
 error_reporting(E_ERROR);
 
-class A2_Optimized_Plugin extends A2_Optimized_LifeCycle {
+class A2_Optimized_Plugin extends A2_Optimized_InstallIndicator {
 
+    const optionInstalled = '_installed';
+    const optionVersion = '_version';
+
+    public function install() {
+
+        // Initialize Plugin Options
+        $this->initOptions();
+
+        // Initialize DB Tables used by the plugin
+        $this->installDatabaseTables();
+
+        // Other Plugin initialization - for the plugin writer to override as needed
+        $this->otherInstall();
+
+        // Record the installed version
+        $this->saveInstalledVersion();
+
+        // To avoid running install() more then once
+        $this->markAsInstalled();
+    }
+
+
+
+    public function uninstall() {
+        $this->otherUninstall();
+        $this->unInstallDatabaseTables();
+        $this->deleteSavedOptions();
+        $this->markAsUnInstalled();
+    }
 
 
     /**
@@ -33,7 +62,7 @@ class A2_Optimized_Plugin extends A2_Optimized_LifeCycle {
         $options = $this->getOptionMetaData();
         if (!empty($options)) {
             foreach ($options as $key => $arr) {
-                if (is_array($arr) && count($arr > 1)) {
+                if (is_array($arr) && count($arr) > 1) {
                     $this->addOption($key, $arr[1]);
                 }
             }
@@ -415,7 +444,7 @@ HTML;
 
     public function myStyleSheet() {
 
-        echo<<<CSS
+        echo<<<STYLE
    <style type="text/css">
         #edge-mode{
             display:none !important;
@@ -427,7 +456,7 @@ HTML;
             display:none;
         }
    </style>
-CSS;
+STYLE;
     }
 
 
@@ -719,6 +748,281 @@ SCRIPT;
 
 
 
+    }
+
+
+    /**
+     * Puts the configuration page in the Plugins menu by default.
+     * Override to put it elsewhere or create a set of submenus
+     * Override with an empty implementation if you don't want a configuration page
+     * @return void
+     */
+    public function addSettingsSubMenuPage() {
+        $this->addSettingsSubMenuPageToMenu();
+    }
+
+
+    protected function requireExtraPluginFiles() {
+        //require_once(ABSPATH . 'wp-includes/pluggable.php');
+        require_once(ABSPATH . 'wp-admin/includes/plugin.php');
+    }
+
+    /**
+     * @return string Slug name for the URL to the Setting page
+     * (i.e. the page for setting options)
+     */
+    protected function getSettingsSlug() {
+        return get_class($this) . '_admin';
+    }
+
+    public function addAdminBar(){
+        $this->requireExtraPluginFiles();
+        global $wp_admin_bar;
+
+        if(current_user_can('manage_options')){
+            $wp_admin_bar->add_node(array(
+                'id'    => 'a2-optimized-admin-bar',
+                'title' => 'A2 Optimized',
+                'href'  => admin_url('admin.php?page='.$this->getSettingsSlug())
+            ));
+        }
+    }
+
+
+    public function addLockedEditor(){
+        $this->requireExtraPluginFiles();
+        add_theme_page( '<span style="color:red !important">Editor Locked</span>', '<span style="color:red !important">Editor Locked</span>', 'manage_options', 'editor-locked', array(&$this,'settingsPage') );
+    }
+
+
+    protected function addSettingsSubMenuPageToPluginsMenu() {
+        $this->requireExtraPluginFiles();
+        $displayName = $this->getPluginDisplayName();
+        add_submenu_page('plugins.php',
+            $displayName,
+            $displayName,
+            'manage_options',
+            $this->getSettingsSlug(),
+            array(&$this, 'settingsPage'));
+    }
+
+    protected function addSettingsSubMenuPageToDashboard() {
+        $this->requireExtraPluginFiles();
+        $displayName = $this->getPluginDisplayName();
+        add_dashboard_page($displayName,
+            $displayName,
+            'manage_options',
+            $this->getSettingsSlug(),
+            array(&$this, 'settingsPage'));
+    }
+
+
+    protected function addSettingsSubMenuPageToMenu() {
+        $this->requireExtraPluginFiles();
+        $displayName = $this->getPluginDisplayName();
+        add_menu_page(   $displayName,
+            $displayName,
+            'manage_options',
+            $this->getSettingsSlug(),
+            array(&$this, 'settingsPage'),null,3.14159265359);
+    }
+
+
+    protected function addSettingsSubMenuPageToSettingsMenu() {
+        $this->requireExtraPluginFiles();
+        $displayName = $this->getPluginDisplayName();
+        add_options_page($displayName,
+            $displayName,
+            'manage_options',
+            $this->getSettingsSlug(),
+            array(&$this, 'settingsPage'));
+    }
+
+
+
+
+    /**
+     * @param  $name string name of a database table
+     * @return string input prefixed with the WordPress DB table prefix
+     * plus the prefix for this plugin (lower-cased) to avoid table name collisions.
+     * The plugin prefix is lower-cases as a best practice that all DB table names are lower case to
+     * avoid issues on some platforms
+     */
+    protected function prefixTableName($name) {
+        global $wpdb;
+        return $wpdb->prefix .  strtolower($this->prefix($name));
+    }
+
+
+    /**
+     * Convenience function for creating AJAX URLs.
+     *
+     * @param $actionName string the name of the ajax action registered in a call like
+     * add_action('wp_ajax_actionName', array(&$this, 'functionName'));
+     *     and/or
+     * add_action('wp_ajax_nopriv_actionName', array(&$this, 'functionName'));
+     *
+     * If have an additional parameters to add to the Ajax call, e.g. an "id" parameter,
+     * you could call this function and append to the returned string like:
+     *    $url = $this->getAjaxUrl('myaction&id=') . urlencode($id);
+     * or more complex:
+     *    $url = sprintf($this->getAjaxUrl('myaction&id=%s&var2=%s&var3=%s'), urlencode($id), urlencode($var2), urlencode($var3));
+     *
+     * @return string URL that can be used in a web page to make an Ajax call to $this->functionName
+     */
+    public function getAjaxUrl($actionName) {
+        return admin_url('admin-ajax.php') . '?action=' . $actionName;
+    }
+
+
+
+
+
+    /**
+     * @return bool indicating if the plugin is installed already
+     */
+    public function isInstalled() {
+        return $this->getOption(self::optionInstalled) == true;
+    }
+
+    /**
+     * Note in DB that the plugin is installed
+     * @return null
+     */
+    protected function markAsInstalled() {
+        return $this->updateOption(self::optionInstalled, true);
+    }
+
+    /**
+     * Note in DB that the plugin is uninstalled
+     * @return bool returned form delete_option.
+     * true implies the plugin was installed at the time of this call,
+     * false implies it was not.
+     */
+    protected function markAsUnInstalled() {
+        return $this->deleteOption(self::optionInstalled);
+    }
+
+    /**
+     * Set a version string in the options. This is useful if you install upgrade and
+     * need to check if an older version was installed to see if you need to do certain
+     * upgrade housekeeping (e.g. changes to DB schema).
+     * @return null
+     */
+    protected function getVersionSaved() {
+        return $this->getOption(self::optionVersion);
+    }
+
+    /**
+     * Set a version string in the options.
+     * need to check if
+     * @param  $version string best practice: use a dot-delimited string like '1.2.3' so version strings can be easily
+     * compared using version_compare (http://php.net/manual/en/function.version-compare.php)
+     * @return null
+     */
+    protected function setVersionSaved($version) {
+        return $this->updateOption(self::optionVersion, $version);
+    }
+
+    /**
+     * Get a value for input key in the header section of main plugin file.
+     * E.g. "Plugin Name", "Version", "Description", "Text Domain", etc.
+     * @param $key string plugin header key
+     * @return string if found, otherwise null
+     */
+    public function getPluginHeaderValue($key) {
+        // Read the string from the comment header of the main plugin file
+        $data = file_get_contents($this->getPluginDir() . DIRECTORY_SEPARATOR . $this->getMainPluginFileName());
+        $match = array();
+        preg_match('/' . $key . ':\s*(\S+)/', $data, $match);
+        if (count($match) >= 1) {
+            return $match[1];
+        }
+        return null;
+    }
+
+    /**
+     * If your subclass of this class lives in a different directory,
+     * override this method with the exact same code. Since __FILE__ will
+     * be different, you will then get the right dir returned.
+     * @return string
+     */
+    protected function getPluginDir() {
+        return dirname(__FILE__);
+    }
+
+    /**
+     * Version of this code.
+     * Best practice: define version strings to be easily compared using version_compare()
+     * (http://php.net/manual/en/function.version-compare.php)
+     * NOTE: You should manually make this match the SVN tag for your main plugin file 'Version' release and 'Stable tag' in readme.txt
+     * @return string
+     */
+    public function getVersion() {
+        return $this->getPluginHeaderValue('Version');
+    }
+
+
+    /**
+     * Useful when checking for upgrades, can tell if the currently installed version is earlier than the
+     * newly installed code. This case indicates that an upgrade has been installed and this is the first time it
+     * has been activated, so any upgrade actions should be taken.
+     * @return bool true if the version saved in the options is earlier than the version declared in getVersion().
+     * true indicates that new code is installed and this is the first time it is activated, so upgrade actions
+     * should be taken. Assumes that version string comparable by version_compare, examples: '1', '1.1', '1.1.1', '2.0', etc.
+     */
+    public function isInstalledCodeAnUpgrade() {
+        return $this->isSavedVersionLessThan($this->getVersion());
+    }
+
+    /**
+     * Used to see if the installed code is an earlier version than the input version
+     * @param  $aVersion string
+     * @return bool true if the saved version is earlier (by natural order) than the input version
+     */
+    public function isSavedVersionLessThan($aVersion) {
+        return $this->isVersionLessThan($this->getVersionSaved(), $aVersion);
+    }
+
+    /**
+     * Used to see if the installed code is the same or earlier than the input version.
+     * Useful when checking for an upgrade. If you haven't specified the number of the newer version yet,
+     * but the last version (installed) was 2.3 (for example) you could check if
+     * For example, $this->isSavedVersionLessThanEqual('2.3') == true indicates that the saved version is not upgraded
+     * past 2.3 yet and therefore you would perform some appropriate upgrade action.
+     * @param  $aVersion string
+     * @return bool true if the saved version is earlier (by natural order) than the input version
+     */
+    public function isSavedVersionLessThanEqual($aVersion) {
+        return $this->isVersionLessThanEqual($this->getVersionSaved(), $aVersion);
+    }
+
+    /**
+     * @param  $version1 string a version string such as '1', '1.1', '1.1.1', '2.0', etc.
+     * @param  $version2 string a version string such as '1', '1.1', '1.1.1', '2.0', etc.
+     * @return bool true if version_compare of $versions1 and $version2 shows $version1 as the same or earlier
+     */
+    public function isVersionLessThanEqual($version1, $version2) {
+        return (version_compare($version1, $version2) <= 0);
+    }
+
+    /**
+     * @param  $version1 string a version string such as '1', '1.1', '1.1.1', '2.0', etc.
+     * @param  $version2 string a version string such as '1', '1.1', '1.1.1', '2.0', etc.
+     * @return bool true if version_compare of $versions1 and $version2 shows $version1 as earlier
+     */
+    public function isVersionLessThan($version1, $version2) {
+        return (version_compare($version1, $version2) < 0);
+    }
+
+    /**
+     * Record the installed version to options.
+     * This helps track was version is installed so when an upgrade is installed, it should call this when finished
+     * upgrading to record the new current version
+     * @return void
+     */
+    protected function saveInstalledVersion() {
+        $this->setVersionSaved($this->getVersion());
     }
 
 }
